ggml: add type_size to data view offset calc
This commit suggests adding the type_size to the calculation of the
offset when creating a view of a tensor.

The motivation for this change is that when I tried calling
`ggml_view_1d` using an offset like this:
```c
    struct ggml_tensor * view = ggml_view_1d(ctx, x, 5, 0);
```
Where the `x` tensor has 10 elments with values ranging from 0 to 9,
and requesting a view with 5 elements, starting at offset zero this
works as expected.

But if I try using an offset of 5, with the expectation (which may be
incorrect) that this would produce a view with the last 5 elements of
the tensor `x`:
```c
    struct ggml_tensor * view = ggml_view_1d(ctx, x, 5, 5);
```
This does not work as I had expected. This commit contains a test with
the above use case, which hopefully clarifies that I'm trying to do and
perhaps someone can point out what I'm doing wrong, or confirm that
this may be an issue with the current implementation.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>