backend: remove duplicate backend typedefs
This commit removes the duplicate typedefs for backend structs in
`ggml-backend.h`. These typedefs are already defined in `ggml-alloc.h`
which `ggml-backend.h` includes.

There might be a case for having the typedefs in `ggml-backend.h` but if
this perhaps is an oversight I think it might better to remove them from
`ggml-backend.h` to avoid confusion.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>