cuda: set backend type to GPU when init tensor
This commit sets the backend type to GPU when initializing a tensor
in the CUDA backend.

The motivation for this change is that currently the backend type of
the tensor is still set to CPU after the tensor is initialized by the
CUDA backend. Other backends like sycl and kompute set the backend
type to GPU, and this change makes the CUDA backend consistent with
those backends.

Signed-off-by: Daniel Bevenius <daniel.bevenius@gmail.com>