SYCL: do not skip CPU if explicitly selected
It can be useful to be able to use SYCL on CPU
e.g. to run tests on a machine without a GPU,
like GH's free runners.

With this commit a user can enable the use
of the CPU with the following env variable:

export ONEAPI_DEVICE_SELECTOR="opencl:cpu"

Signed-off-by: Salvatore Mesoraca <s.mesoraca16@gmail.com>