Add option not to abort on cuda OOM
Warning: Not ready for merge.
Add option not to abort on cuda OOM but throw/return a ggml_status.
The goal in this ticket is NOT to be able to continue inference when
OOM, but just to do a clean controlled exit at higher level.
No change to default behavior (abort).
Retouch ggml_tallocr_alloc to return a ggml_status.
Ass a new (cuda build only) unit test to check the no abort flow (skiped
if the envvar
GGML_CUDA_NO_ABORT is not set).